% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{std_search}
\alias{std_search}
\title{Search api for stdict korean}
\usage{
std_search(query, start = 1, count = 10)
}
\arguments{
\item{query}{query for search}

\item{start}{query result of start}

\item{count}{count of result number}
}
\description{
stdict korean api can use search api
}
