% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{std_view}
\alias{std_view}
\title{View api for stdict korean}
\usage{
std_view(query, method = "word_info")
}
\arguments{
\item{query}{query for view}

\item{method}{method. default is "word_info"}
}
\description{
stdict korean api can use view api
}
