% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdictapi-package.R
\docType{package}
\name{stdictapi-package}
\alias{stdictapi}
\alias{stdictapi-package}
\title{stdictapi: Data Package using Korean Standard Dictionary}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

stdictapi is a client package that uses the API provided by the Korean standard dictionary<https://stdict.korean.go.kr/openapi/openApiInfo.do>. You can also directly download a dictionary and use it as a tibble data type.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrchypark.github.io/stdictapi/}
  \item Report bugs at \url{https://github.com/mrchypark/stdictapi/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
