% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdictkr-package.R
\docType{package}
\name{stdictkr-package}
\alias{stdictkr}
\alias{stdictkr-package}
\title{stdictkr: Data Package using Korean Standard Dictionary}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

stdictkr is a client package that uses the API provided by 'the Korean standard dictionary' \url{https://stdict.korean.go.kr/openapi/openApiInfo.do}. You can also directly download a dictionary and use it as a tibble data type.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrchypark.github.io/stdictkr/}
  \item Report bugs at \url{https://github.com/mrchypark/stdictkr/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
